pyproject.toml
[project]
name = "lean-ai"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=7.1.0",
    "langchain-core>=1.2.0",
    "langchain-openai>=1.1.3",
    "langgraph>=1.0.5",
    "langgraph-checkpoint>=3.0.1",
    "langgraph-prebuilt>=1.0.5",
    "langgraph-sdk>=0.3.0",
    "langsmith>=0.4.59",
    "python-dotenv>=1.2.1",
    "pydantic>=2.12.5",
]
